backend/app/requirements.txt
# fastapi==0.95.2
# uvicorn[standard]==0.22.0
# SQLAlchemy==1.4.52
# psycopg2-binary==2.9.7
# python-jose==3.3.0
# passlib[bcrypt]==1.7.4
# pydantic==1.10.9
# requests==2.31.0
# pdfkit==1.0.0
# python-multipart==0.0.6
# alembic==1.11.1
# python-dotenv==1.0.0


# fastapi==0.95.2
# uvicorn[standard]==0.22.0
# SQLAlchemy==1.4.52
# psycopg[binary,pool]      # PostgreSQL driver (works with Python 3.13)
# python-jose==3.3.0
# passlib[bcrypt]==1.7.4
# pydantic==1.10.9
# requests==2.31.0
# pdfkit==1.0.0
# python-multipart==0.0.6
# alembic==1.11.1
# python-dotenv==1.0.0

fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.52
psycopg2-binary==2.9.7
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.9
requests==2.31.0
pdfkit==1.0.0
python-multipart==0.0.6
alembic==1.11.1
python-dotenv==1.0.0
twilio
pydantic[email]